backend/requirements.txt
pymongo==4.6.0
python-dotenv==1.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
ultralytics>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0
requests==2.31.0